Switch between multi and single person
Added a switch between two modes. When detecting multiple people, the blinking rate resets if we lose focus of the person. In order to work around that issue, I have added a switch, so when we know we have only one person on the screen, there is no reason for the blinking rate to keep resetting.